Refactor SpanContextPropagatorExtensions into a helper class because the object on the extension method is never used
